src/main/java/jp/lg/tokyo/metro/mansion/todokede/common/annotation/validator/HaftSizeNumberValidator.java
/*
 * @(#) HaftSizeNumberValidator.java
 *
 * Copyright(C) 2019 by Hitachi Information Systems CO., LTD
 *
 * Description:
 * @author ptluan
 * Create Date: 2019/11/25
 * Version: 1.0
 */

package jp.lg.tokyo.metro.mansion.todokede.common.annotation.validator;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;

import org.apache.commons.lang.StringUtils;

import jp.lg.tokyo.metro.mansion.todokede.common.annotation.HaftSizeNumber;
import jp.lg.tokyo.metro.mansion.todokede.common.utils.CommonUtils;

public class HaftSizeNumberValidator implements ConstraintValidator<HaftSizeNumber, String> {

    @Override
    public void initialize(HaftSizeNumber constraintAnnotation) {
    }

    @Override
    public boolean isValid(String input, ConstraintValidatorContext context) {
        if (StringUtils.isBlank(input)) {
            return true;
        }
        /* returns true if all input characters are half size number */
        return CommonUtils.isHalfSizeRatio(input);
    }

}
